Moved semantic_map_publisher headers into include/semantic_map_publisher
