+++ semantic_map_publisher/src/semantic_map_publisher.cpp
@@ -1,1 +1,1 @@
-#include "semantic_map_publisher.h"
+#include "semantic_map_publisher/semantic_map_publisher.h"
