+++ semantic_map_publisher/src/semantic_map_publisher_main.cpp
@@ -1,4 +1,4 @@
-#include "semantic_map_publisher.h"
+#include "semantic_map_publisher/semantic_map_publisher.h"
 
 int main(int argc, char **argv)
 {
